requirements.txt
uagents==0.21.0
aiohttp==3.9.1
requests>=2.32.3
python-dotenv==1.0.0
cosmpy>=0.9.2 